--- client/src/features/coverletter/index.ts
@@ -1,1 +1,0 @@
-export { default as CoverLetter } from './CoverLetter';
