--- client/src/features/coverletter/list/index.ts
@@ -1,2 +1,0 @@
-export { default as CoverLetterList } from './CoverLetterItems';
-export { default as CoverLetterItems } from './CoverLetterItems';
